requirements.txt
black==22.12.0
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
db-dtypes==1.0.5
google-api-core==2.11.0
google-auth==2.15.0
google-auth-oauthlib==0.8.0
google-cloud-bigquery==3.4.1
google-cloud-bigquery-storage==2.17.0
google-cloud-core==2.3.2
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
grpcio==1.51.1
grpcio-status==1.51.1
idna==3.4
mypy-extensions==0.4.3
numpy==1.23.5
oauthlib==3.2.2
packaging==21.3
pandas==1.5.2
pandas-gbq==0.18.1
pathspec==0.10.3
platformdirs==2.6.0
proto-plus==1.22.1
protobuf==4.21.12
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydata-google-auth==1.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tomli==2.0.1
urllib3==1.26.13
